requirements.txt
# Web Framework
streamlit>=1.28.0
streamlit-autorefresh>=1.0.1

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# News API
requests>=2.31.0
newsapi-python>=0.2.7

# OpenAI / OpenRouter
openai>=1.0.0

# Article Scraping
newspaper3k>=0.2.8
lxml_html_clean>=0.1.0

# Environment Management
python-dotenv>=1.0.0

# NLP Utilities (Lightweight)
nltk>=3.8.0
